+++ xdclass-user-service/src/main/java/net/xdclass/service/AddressService.java
@@ -13,6 +13,6 @@ import com.baomidou.mybatisplus.extension.service.IService;
  * @createTime 2024/10/4 16:58
  */
 public interface AddressService extends IService<AddressDO> {
-    AddressDO detail(int id);
+    AddressDO detail(long id);
 
 }
